crypto: load the OpenSSL 3 legacy provider so username-hash=md4 works
extra/chap-secrets.c resolves the "username-hash" option through
EVP_get_digestbyname().  On OpenSSL 3.0 and later MD4 was moved into the
legacy provider, which is not activated by default, so EVP_DigestInit_ex()
fails for md4 and the hash chain silently produces nothing -- on every
current distro.  The low-level MD4_Init()/MD4_Update() path still works, so
MSCHAP is unaffected; only the EVP consumer breaks.

This is a pre-existing bug, not something introduced by the backend work --
it just became visible because the new known-answer tests exercise every
digest through both the direct and the EVP interface, and md4 failed on the
OPENSSL backend while passing on TOMCRYPT and INTERNAL.

Add ap_crypto_init(), called once from main(), which activates the legacy
provider on OpenSSL 3.x.  The default provider is activated explicitly too:
OpenSSL only auto-activates it while no provider has been loaded, so loading
legacy on its own would take every other digest down with it.

The libtomcrypt backends get the same entry point; there it just runs the
hash registration eagerly instead of leaving it to the lazy path.
